# openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: SPA Development Reference API
  version: 0.1.0
  description: |
    Normative browser-facing HTTP contract of the SPA development reference
    application, a generic request and approval sample.

    All business endpoints require a Bearer access token. In the AWS
    deployment (`APP_MODE=aws`) it is a JWT access token issued by Amazon
    Cognito and the application role of the caller (Requester, Approver, or
    Administrator) is derived from the `cognito:groups` claim of the validated
    token. In the local demo (`APP_MODE=local`) it is one of three fixed demo
    tokens and the backend maps each to one role. In both modes the roles are
    derived by the server, and each operation states the roles it requires.
    A caller that lacks the required role, or that is not permitted to use the
    addressed resource, receives `403 FORBIDDEN`. The Administrator role is not
    a superuser.

    Evaluation order for a business operation: authentication, coarse role,
    structural validation (path, query, content type, body), resource lookup,
    resource authorization, version, state. The first failing step determines
    the response; in particular, when a supplied `version` is stale and the
    current state is also invalid, the response is `409 CONCURRENCY_CONFLICT`.
    An unknown path is `404 ROUTE_NOT_FOUND` and an unsupported method on a known
    path is `405 METHOD_NOT_ALLOWED`; both are Problem responses. The detailed
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/requests:
    get:
      operationId: listRequests
      tags: [Requests]
      summary: List requests visible to the caller
      description: |
        Requester: only requests whose `requesterId` is the caller. Administrator:
        all requests. A caller with both roles sees the union. A caller with
        neither role (for example an Approver only) receives `403 FORBIDDEN`.

        Order: `updatedAt` descending, then `id` ascending. Pagination is
        applied after the visibility filter.
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Requests visible to the caller.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/IdentityProviderUnavailable'
    post:
      operationId: createRequest
      tags: [Requests]
      summary: Create a request in DRAFT state
      description: |
        Requires the Requester role and a verified email address on the caller's
        identity; otherwise `403 FORBIDDEN`. The `requesterId` of the created
        request is the caller's subject. If the identity provider's UserInfo
        endpoint is unavailable the response is
        `503 IDENTITY_PROVIDER_UNAVAILABLE`. The title is trimmed before
        validation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRequestInput'
      responses:
        '201':
          description: The request was created in `DRAFT` state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Request'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/IdentityProviderUnavailable'
  /api/requests/{requestId}:
    parameters:
      - $ref: '#/components/parameters/RequestId'
    get:
      operationId: getRequest
      tags: [Requests]
      summary: Get a request
      description: |
        Requester: own request in any state. Approver: requests in `SUBMITTED`
        state only. Administrator: any request. Otherwise `403 FORBIDDEN`. A
        request that does not exist is `404`.
      responses:
        '200':
          description: The request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Request'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/IdentityProviderUnavailable'
    put:
      operationId: updateDraftRequest
      tags: [Requests]
      summary: Update a draft request
      description: |
        Requires the Requester role and ownership of the request (otherwise
        `403 FORBIDDEN`). A stale `version` is `409 CONCURRENCY_CONFLICT`, and
        this takes precedence over a state error. With a matching version, only a
        request in `DRAFT` state can be updated; otherwise
        `409 REQUEST_INVALID_STATE`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDraftRequestInput'
      responses:
        '200':
          description: The updated request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Request'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/IdentityProviderUnavailable'
  /api/requests/{requestId}/submit:
    parameters:
      - $ref: '#/components/parameters/RequestId'
    post:
      operationId: submitRequest
      tags: [Requests]
      summary: Submit a draft request
      description: |
        Transitions the request from `DRAFT` to `SUBMITTED`. Requires the
        Requester role and ownership of the request (otherwise
        `403 FORBIDDEN`). A stale `version` is `409 CONCURRENCY_CONFLICT`, and
        this takes precedence over a state error. With a matching version, a
        request not in `DRAFT` state is `409 REQUEST_INVALID_STATE`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VersionCommand'
      responses:
        '200':
          description: The submitted request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Request'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/IdentityProviderUnavailable'
  /api/requests/{requestId}/approve:
    parameters:
      - $ref: '#/components/parameters/RequestId'
    post:
      operationId: approveRequest
      tags: [Approvals]
      summary: Approve a submitted request
      description: |
        Transitions the request from `SUBMITTED` to `APPROVED`. Requires the
        Approver role (otherwise `403 FORBIDDEN`). A stale `version` is
        `409 CONCURRENCY_CONFLICT`, and this takes precedence over a state
        error. With a matching version, a request not in `SUBMITTED` state is
        `409 REQUEST_INVALID_STATE`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DecisionCommand'
      responses:
        '200':
          description: The approved request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Request'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/IdentityProviderUnavailable'
  /api/requests/{requestId}/reject:
    parameters:
      - $ref: '#/components/parameters/RequestId'
    post:
      operationId: rejectRequest
      tags: [Approvals]
      summary: Reject a submitted request
      description: |
        Transitions the request from `SUBMITTED` to `REJECTED`. Requires the
        Approver role (otherwise `403 FORBIDDEN`). A stale `version` is
        `409 CONCURRENCY_CONFLICT`, and this takes precedence over a state
        error. With a matching version, a request not in `SUBMITTED` state is
        `409 REQUEST_INVALID_STATE`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DecisionCommand'
      responses:
        '200':
          description: The rejected request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Request'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/IdentityProviderUnavailable'
  /api/approvals:
    get:
      operationId: listApprovalQueue
      tags: [Approvals]
      summary: List requests awaiting a decision
      description: |
        Requires the Approver role (otherwise `403 FORBIDDEN`). Returns requests
        in `SUBMITTED` state.

        Order: `updatedAt` ascending, then `id` ascending, so that the oldest
        submitted request comes first. Pagination is applied after the state
        filter.
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Requests awaiting a decision.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/IdentityProviderUnavailable'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        AWS deployment: Amazon Cognito access token only (an ID token is not
        accepted as authorization evidence). The BFF validates the token and
        derives the application roles from its `cognito:groups` claim.
        Local demo (`APP_MODE=local`): one of the fixed, non-secret demo tokens,
        which the AWS deployment never accepts. In both modes role values
        supplied by the browser are never used.
  parameters:
    RequestId:
      name: requestId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Limit:
      name: limit
      in: query
      required: false
      description: Maximum number of items in the page.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    Cursor:
      name: cursor
      in: query
      required: false
      description: |
        Opaque cursor taken from `nextCursor` of the previous page. Clients do
        not interpret its content. Cursors do not expire. A cursor may be bound to
        endpoint-specific filters and resource identifiers: the audit list binds
        it to the `requestId` filter, and the attachment list binds it to the
        `requestId` path value. A malformed cursor, or a cursor used with a
        different endpoint or a different bound value, is a
        `400 VALIDATION_ERROR`. Pages are not
        taken from a database snapshot: concurrent changes between page requests
        can move or skip items.
      schema:
        type: string
        minLength: 1
  responses:
    BadRequest:
      description: Validation failure (`VALIDATION_ERROR`).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            title: Validation error
            status: 400
            detail: The title must be between 1 and 200 characters.
            code: VALIDATION_ERROR
            traceId: example-trace-id
    Unauthorized:
      description: Authentication is required or the token is invalid (`AUTHENTICATION_REQUIRED`).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            title: Authentication required
            status: 401
            code: AUTHENTICATION_REQUIRED
            traceId: example-trace-id
    Forbidden:
      description: The caller is not permitted to perform the operation (`FORBIDDEN`).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            title: Forbidden
            status: 403
            code: FORBIDDEN
            traceId: example-trace-id
    NotFound:
      description: The request or attachment does not exist (`REQUEST_NOT_FOUND` or `ATTACHMENT_NOT_FOUND`).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            title: Request not found
            status: 404
            code: REQUEST_NOT_FOUND
            traceId: example-trace-id
    Conflict:
      description: The supplied version is stale (`CONCURRENCY_CONFLICT`, which takes precedence) or, with a matching version, the request is not in a valid state for the operation (`REQUEST_INVALID_STATE`).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            title: Invalid state transition
            status: 409
            detail: The request is not in a state that can be approved.
            code: REQUEST_INVALID_STATE
            traceId: example-trace-id
...
    UnsupportedMediaType:
      description: The media type of the request body or uploaded file is not supported, or the file content or extension does not match it (`UNSUPPORTED_MEDIA_TYPE`).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            title: Unsupported media type
            status: 415
            code: UNSUPPORTED_MEDIA_TYPE
            traceId: example-trace-id
    IdentityProviderUnavailable:
      description: The identity provider needed to verify the token or the caller's email is unavailable (`IDENTITY_PROVIDER_UNAVAILABLE`).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            title: Identity provider unavailable
            status: 503
            code: IDENTITY_PROVIDER_UNAVAILABLE
            traceId: example-trace-id
...
    InternalError:
      description: Unexpected implementation, database, or unclassified infrastructure failure (`INTERNAL_ERROR`).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            title: Internal error
            status: 500
            code: INTERNAL_ERROR
            traceId: example-trace-id
  schemas:
...
    RequestStatus:
      type: string
      enum: [DRAFT, SUBMITTED, APPROVED, REJECTED]
    Request:
      type: object
      required: [id, requesterId, title, description, status, version, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        requesterId:
          type: string
          minLength: 1
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          minLength: 0
          maxLength: 5000
        status:
          $ref: '#/components/schemas/RequestStatus'
        version:
          type: integer
          minimum: 1
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    RequestSummary:
      description: Summary of a request. It carries the same fields as `Request`.
      allOf:
        - $ref: '#/components/schemas/Request'
    RequestList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/RequestSummary'
        nextCursor:
          type: string
          minLength: 1
          description: Cursor for the next page. Omitted when there is no next page.
    CreateRequestInput:
      type: object
      required: [title, description]
      additionalProperties: false
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          minLength: 0
          maxLength: 5000
    UpdateDraftRequestInput:
      type: object
      required: [title, description, version]
      additionalProperties: false
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          minLength: 0
          maxLength: 5000
        version:
          type: integer
          minimum: 1
    VersionCommand:
      type: object
      required: [version]
      additionalProperties: false
      properties:
        version:
          type: integer
          minimum: 1
    DecisionCommand:
      type: object
      required: [version]
      additionalProperties: false
      properties:
        version:
          type: integer
          minimum: 1
        comment:
          type: string
          maxLength: 2000
...
    ErrorCode:
      description: Stable machine-readable application error identifier.
      type: string
      enum:
        - VALIDATION_ERROR
        - AUTHENTICATION_REQUIRED
        - FORBIDDEN
        - REQUEST_NOT_FOUND
        - ATTACHMENT_NOT_FOUND
        - ROUTE_NOT_FOUND
        - METHOD_NOT_ALLOWED
        - REQUEST_INVALID_STATE
        - CONCURRENCY_CONFLICT
        - PAYLOAD_TOO_LARGE
        - UNSUPPORTED_MEDIA_TYPE
        - IDENTITY_PROVIDER_UNAVAILABLE
        - OBJECT_STORAGE_UNAVAILABLE
        - INTERNAL_ERROR
    Problem:
      description: |
        Problem Details (`application/problem+json`) with stable application
        extension fields. Status mapping: `VALIDATION_ERROR` 400,
        `AUTHENTICATION_REQUIRED` 401, `FORBIDDEN` 403, `REQUEST_NOT_FOUND` and
        `ATTACHMENT_NOT_FOUND` 404, `ROUTE_NOT_FOUND` 404,
        `METHOD_NOT_ALLOWED` 405, `REQUEST_INVALID_STATE` and
        `CONCURRENCY_CONFLICT` 409, `PAYLOAD_TOO_LARGE` 413,
        `UNSUPPORTED_MEDIA_TYPE` 415, `IDENTITY_PROVIDER_UNAVAILABLE` and
        `OBJECT_STORAGE_UNAVAILABLE` 503, `INTERNAL_ERROR` 500.
      type: object
      required: [title, status, code, traceId]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        code:
          $ref: '#/components/schemas/ErrorCode'
        traceId:
          type: string